--- a/mfo/admin/services/syllabus.py
+++ b/mfo/admin/services/syllabus.py
@@ -2,11 +2,11 @@ import pandas as pd
 import flask
 import tabula
 import sqlalchemy as sa
+import traceback
 
 from mfo.database.base import db
-from mfo.database.models import FestivalClass
+from mfo.database.models import FestivalClass, DefaultTimes, Discipline, Level, PerformanceType
 from mfo.admin.services.admin_services import infer_attributes
-
 
 def verify_pdf(tables):
     """Verify that the tables extracted from the PDF file are in the expected format."""
@@ -87,6 +87,7 @@ def load_database(combined_df):
     try:
         # find all rows in combined_df that have the same number
         grouped_df = combined_df.groupby('number')
+
         for number, group in grouped_df:
             # get list of classes with the same number from the database
             stmt = sa.select(FestivalClass).where(FestivalClass.number == number)
@@ -95,8 +96,49 @@ def load_database(combined_df):
             # add all the rows in the group
             if not db_classes:
                 for index, row in group.iterrows():
-                    suffix, class_type, discipline = infer_attributes(row)
+                    suffix, class_type, discipline, level = infer_attributes(row)
+                    
+                    # get adjudication time from the database
+
                     if (row.number, suffix) not in existing_classes:
+
+                        discipline_subquery = (
+                            sa.select(Discipline.id)
+                            .where(sa.func.lower(Discipline.name) == discipline.lower())
+                        ).subquery()
+
+                        level_subquery = (
+                            sa.select(Level.id)
+                            .where(sa.func.lower(Level.name) == level.lower())
+                        ).subquery()
+
+                        type_subquery = (
+                            sa.select(PerformanceType.id)
+                            .where(sa.func.lower(PerformanceType.name) == class_type.lower())
+                        ).subquery()
+
+                        adjudication_time = db.session.execute(
+                            sa.select(DefaultTimes.time)
+                            .where(
+                                sa.and_(
+                                    DefaultTimes.time_type == "adjudication",
+                                    DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                    DefaultTimes.level_id == level_subquery.c.id
+                                )
+                            )
+                        ).scalar()
+
+                        move_time = db.session.execute(
+                            sa.select(DefaultTimes.time)
+                            .where(
+                                sa.and_(
+                                    DefaultTimes.time_type == "move",
+                                    DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                    DefaultTimes.performance_type_id == type_subquery.c.id
+                                )
+                            )
+                        ).scalar()
+                        
                         festival_class = FestivalClass(
                             number=row.number, 
                             suffix=suffix, 
@@ -104,11 +146,15 @@ def load_database(combined_df):
                             class_type=class_type,
                             discipline=discipline,
                             fee=row.fee,
-                            adjudication_time = flask.current_app.config["DEFAULT_ADJUDICATION_TIME"][discipline][class_type],
-                            move_time = flask.current_app.config["DEFAULT_MOVE_TIME"][discipline][class_type],
+                            level=level,
+                            adjudication_time = adjudication_time,
+                            move_time = move_time,
                             )
 
                         db.session.add(festival_class)
+
+                        print(f"Added class {row.number}{suffix}   {row.description}  {class_type}  {discipline}  {level}")
+
                     existing_classes.append((row.number, suffix))
                     
             else:
@@ -116,9 +162,9 @@ def load_database(combined_df):
                 # update existing classes in the database with information from the dataframe
                 for db_class in db_classes:
                     for index, row in group.iterrows():
-                        suffix, class_type, discipline = infer_attributes(row)
+                        suffix, class_type, discipline, level = infer_attributes(row)
                         if (row.number, suffix) not in existing_classes:
-                            if (db_class.number, db_class.suffix) == (row.number, suffix):
+                            if (db_class.number, db_class.suffix) == (row.number, suffix):                        
                                 db_class.name = row.description
                                 db_class.class_type = class_type
                                 db_class.discipline = discipline
@@ -126,19 +172,66 @@ def load_database(combined_df):
 
                                 # overwrite adjudication_time and move_time with defaults, if none exist
                                 if pd.isna(db_class.adjudication_time):
-                                    db_class.adjudication_time = flask.current_app.config["DEFAULT_ADJUDICATION_TIME"][discipline][class_type]
+
+                                    discipline_subquery = (
+                                        sa.select(Discipline.id)
+                                        .where(sa.func.lower(Discipline.name) == discipline.lower())
+                                    ).subquery()
+
+                                    level_subquery = (
+                                        sa.select(Level.id)
+                                        .where(sa.func.lower(Level.name) == level.lower())
+                                    ).subquery()
+
+                                    adjudication_time = db.session.execute(
+                                        sa.select(DefaultTimes.time)
+                                        .where(
+                                            sa.and_(
+                                                DefaultTimes.time_type == "adjudication",
+                                                DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                                DefaultTimes.level_id == level_subquery.c.id
+                                            )
+                                        )
+                                    ).scalar()
+
+                                    db_class.adjudication_time = adjudication_time
                                 
                                 if pd.isna(db_class.move_time):
-                                    db_class.move_time = flask.current_app.config["DEFAULT_MOVE_TIME"][discipline][class_type]
+
+                                    discipline_subquery = (
+                                        sa.select(Discipline.id)
+                                        .where(sa.func.lower(Discipline.name) == discipline.lower())
+                                    ).subquery()
+
+                                    type_subquery = (
+                                        sa.select(PerformanceType.id)
+                                        .where(sa.func.lower(PerformanceType.name) == class_type.lower())
+                                    ).subquery()
+
+                                    move_time = db.session.execute(
+                                        sa.select(DefaultTimes.time)
+                                        .where(
+                                            sa.and_(
+                                                DefaultTimes.time_type == "move",
+                                                DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                                DefaultTimes.performance_type_id == type_subquery.c.id
+                                            )
+                                        )
+                                    ).scalar()
+
+                                    db_class.move_time = move_time
 
                                 db.session.add(db_class)
+
+                                print(f"Added class {row.number}{suffix}   {row.description}  {class_type}  {discipline}  {level}")
+ 
                                 existing_classes.append((row.number, suffix))
 
                 # Update classes from the database that are not in the dataframe group, 
                 # using the fields from the first row in the dataframe group
                 for db_class in db_classes:
                     row = group.iloc[0]
-                    suffix, class_type, discipline = infer_attributes(row)
+                    suffix, class_type, discipline, level = infer_attributes(row)
                     if (db_class.number, db_class.suffix) not in existing_classes:
                         db_class.name = row.description
                         db_class.class_type = class_type
@@ -146,12 +239,58 @@ def load_database(combined_df):
                         db_class.fee = row.fee
 
                         if pd.isna(db_class.adjudication_time):
-                            db_class.adjudication_time = flask.current_app.config["DEFAULT_ADJUDICATION_TIME"][discipline][class_type]
-                                
+
+                            discipline_subquery = (
+                                sa.select(Discipline.id)
+                                .where(sa.func.lower(Discipline.name) == discipline.lower())
+                            ).subquery()
+
+                            level_subquery = (
+                                sa.select(Level.id)
+                                .where(sa.func.lower(Level.name) == level.lower())
+                            ).subquery()
+
+                            adjudication_time = db.session.execute(
+                                sa.select(DefaultTimes.time)
+                                .where(
+                                    sa.and_(
+                                        DefaultTimes.time_type == "adjudication",
+                                        DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                        DefaultTimes.level_id == level_subquery.c.id
+                                    )
+                                )
+                            ).scalar()
+
+                            db_class.adjudication_time = adjudication_time
+                        
                         if pd.isna(db_class.move_time):
-                            db_class.move_time = flask.current_app.config["DEFAULT_MOVE_TIME"][discipline][class_type]
+                            discipline_subquery = (
+                                sa.select(Discipline.id)
+                                .where(sa.func.lower(Discipline.name) == discipline.lower())
+                            ).subquery()
+
+                            type_subquery = (
+                                sa.select(PerformanceType.id)
+                                .where(sa.func.lower(PerformanceType.name) == class_type.lower())
+                            ).subquery()
+
+                            move_time = db.session.execute(
+                                sa.select(DefaultTimes.time)
+                                .where(
+                                    sa.and_(
+                                        DefaultTimes.time_type == "move",
+                                        DefaultTimes.discipline_id == discipline_subquery.c.id,
+                                        DefaultTimes.performance_type_id == type_subquery.c.id
+                                    )
+                                )
+                            ).scalar()
+
+                            db_class.move_time = move_time
 
                         db.session.add(db_class)
+
+                        print(f"Added class {row.number}{suffix}   {row.description}  {class_type}  {discipline}  {level}")
+
                         existing_classes.append((row.number, suffix))
 
         flask.flash("Syllabus added to database. Now you may upload the entries spreadsheet", "success")
@@ -160,6 +299,7 @@ def load_database(combined_df):
     except Exception as e:
         db.session.rollback()
         flask.flash(f"Error adding Syllabus to database: {e}. Database not updated.", "danger")
+        print(traceback.format_exc())
         # issues.append(f"**** Error adding Syllabus to database: {e}")