--- a/tests/qcmf_data.json
+++ b/tests/qcmf_data.json
@@ -223,6 +223,7 @@
         "Composition"
     ],
     "LEVELS": [
+        "None",
         "Pre-Junior",
         "Junior",
         "Intermediate",