## Current state

The *FestivalClass* database model has *adjudication_time* and *move_time* columns for each class. The default times used to populate the class data is currently stored as dictionaries in the *config.py* module. But, its structure is different than that recently proposed by the client. Currently, we align the different default times with the discipline (brass, etc...) and type (solo. etc...). But, the client has standard times per discipline, type, and level (junior, etc...). The problem is that, in the Entries spreadsheet, there is no useful information from which we can derive the class level.

There are over 2,000 classes so asking the client to manually update them all is not realistic.


### Default times provided by client

**Strings**

* Pre-Junior: 5 minutes
* Junior: 5 minutes
* Intermediate: 6 minutes
* Senior: 8 minutes
* Advanced: 9 minutes

**All other disciplines**

* Pre-Junior: 4 minutes
* Junior: 5 minutes
* intermediate: 6 minutes
* Senior: 7 minutes
* Advanced: 8 minutes
* Choirs (all levels): 20 - 25 minutes for each group and that covers their pieces as well
* Bands (all levels): 25 minutes and that covers the repertoire timing.
* Ensembles (all levels): 7 minutes adjudication time + the repertoire time
* I assume Orchestras are treated the same as Bands.

#### Mapping RCM "grades" to levels:

* Pre-Junior: Preparatory to Grade 2
* Junior: Grade 2 to Grade 5
* Intermediate: Grade 5 to Grade 8
* Senior: Grade 8 to 10
* Advanced: Grade 10 to ARCT

#### Mapping ages to levels

* Pre-Junior: 8 years and under
* Junior: 11 years and under
* intermediate: 14 years and under
* Senior: 18 years and under
* advanced: 30 years and under

#### Examples of class names

* Violin Solo - Sonata - One Movement - Junior
* Traditional String Solo - Violin - Junior - 11 years and under
* Violin - Graded Classes - Grade 3 - List A
* Viola Solo - 9 years and under - First year of study
* Viola Solo - Concerto - One movement - 17 years and under
* Intermediate School String **Orchestra**
* Recorder - Folk Song Class - Under four years of instruction
* Harp Solo, List A - RCM Grade 6
* Baritone Saxophone Solo - School Music - Level 3
* Elementary School **Choir** - Male Unison - Grades 1-6
* Band Classes - Intermediate **Band** - Less than 3 years of instruction
* School Music - Mixed Instruments - Small **Ensembles** - Level 4


## Data

In the short term, I need only create functionality that supports deriving levels from the class titles in the Syllabus PDF, then assigning adjudication times to them according to the discipline, type, and level defaults outlined by the client.

When parsing the spreadsheet, check the title of each piece.

Look for the following keys and assign a level accordingly. 

Go through the titles first and assign adjudication times. 

If "choir", "orchestra", "band", or "ensemble" are in the title, set the adjudication time according to the criteria for large groups. Time is expressed in seconds.

| Key | Level | Adjudication Time |
|-----|-------|-------------------|
| choir | * | 1200 |
| chorus | * | 1200 |
| choral concert group | * | 1200 |
| band | * | 1200 |
| ensemble | * | 420 |
| orchestra | * | 1200 |

Then, evaluate the level of the class, according to the following criteria, in order. **Choose the first match**.

| Key | Level |
|-----|-------|
| pre-junior | pre-junior |
| junior | junior |
| intermediate | intermediate |
| senior | senior |
| advanced | advanced |
| preparatory | pre-junior |
| grade 1 | pre-junior |
| grade 2 | pre-junior |
| grade 3 | junior |
| grade 4 | junior |
| grade 5 | intermediate |
| grade 6 | intermediate |
| grade 7 | intermediate |
| grade 8 | senior |
| grade 9 | senior |
| grade 10 | advanced |
| artc | advanced |
| level 1 | pre-junior |
| level 2 | pre-junior |
| level 3 | junior |
| level 4 | junior |
| level 5 | intermediate |
| level 6 | intermediate |
| level 7 | intermediate |
| level 8 | senior |
| level 9 | senior |
| level 10 | advanced |
| first year of study | pre-junior |
| under one year of instruction | pre-junior |
| under two years of instruction | pre-junior |
| under three years of instruction | junior |
| under four years of instruction | junior |
| over four years of instruction | intermediate |
| under five years of instruction | intermediate |
| over five years of instruction | intermediate |
| under six years of instruction | intermediate |
| over six years of instruction | senior |
| primary school | pre-junior |
| elementary school | junior |
| intermediate high school | intermediate |
| senior high school | senior |
| grades k-3 | pre-junior |
| k-3 | pre-junior |
| grades 4-6 | junior |
| grades k-6 | junior |
| grades 3-4 | junior |
| grades 5-6 | junior |
| 5 years and under | pre-junior |
| 6 years and under | pre-junior |
| 7 years and under | pre-junior |
| 8 years and under | pre-junior |
| 9 years and under | junior |
| 10 years and under | junior |
| 11 years and under | junior |
| 12 years and under | intermediate |
| 13 years and under | intermediate |
| 14 years and under | intermediate |
| 15 years and under | senior |
| 16 years and under | senior |
| 17 years and under | senior |
| 18 years and under | senior |
| 19 years and under | advanced |
| 20 years and under | advanced |
| 21 years and under | advanced |
| 22 years and under | advanced |
| 23 years and under | advanced |
| 24 years and under | advanced |
| 25 years and under | advanced |
| 26 years and under | advanced |
| 27 years and under | advanced |
| 28 years and under | advanced |
| 29 years and under | advanced |
| 30 years and under | advanced |

Then, according to discipline and level, set the adjudication time, if it has not yet been set in the first step.

| discipline | level | adjudication time | move time |
|------------|-------|-------------------|-----------|
| none | pre-junior | 240 | 0 |
| none | junior | 300 | 0 |
| none | intermediate | 360 | 0 |
| none | senior | 420 | 0 |
| none | advanced | 480 | 0 |
| vocal | pre-junior | 240 | 0 |
| vocal | junior | 300 | 0 |
| vocal | intermediate | 360 | 0 |
| vocal | senior | 420 | 0 |
| vocal | advanced | 480 | 0 |
| piano | pre-junior | 240 | 0 |
| piano | junior | 300 | 0 |
| piano | intermediate | 360 | 0 |
| piano | senior | 420 | 0 |
| piano | advanced | 480 | 0 |
| organ | pre-junior | 240 | 0 |
| organ | junior | 300 | 0 |
| organ | intermediate | 360 | 0 |
| organ | senior | 420 | 0 |
| organ | advanced | 480 | 0 |
| strings | pre-junior | 300 | 0 |
| strings | junior | 300 | 0 |
| strings | intermediate | 360 | 0 |
| strings | senior | 480 | 0 |
| strings | advanced | 540 | 0 |
| recorder | pre-junior | 240 | 0 |
| recorder | junior | 300 | 0 |
| recorder | intermediate | 360 | 0 |
| recorder | senior | 420 | 0 |
| recorder | advanced | 480 | 0 |
| woodwinds | pre-junior | 240 | 0 |
| woodwinds | junior | 300 | 0 |
| woodwinds | intermediate | 360 | 0 |
| woodwinds | senior | 420 | 0 |
| woodwinds | advanced | 480 | 0 |
| brass | pre-junior | 240 | 0 |
| brass | junior | 300 | 0 |
| brass | intermediate | 360 | 0 |
| brass | senior | 420 | 0 |
| brass | advanced | 480 | 0 |
| percussion | pre-junior | 240 | 0 |
| percussion | junior | 300 | 0 |
| percussion | intermediate | 360 | 0 |
| percussion | senior | 420 | 0 |
| percussion | advanced | 480 | 0 |
| instrumental | pre-junior | 240 | 0 |
| instrumental | junior | 300 | 0 |
| instrumental | intermediate | 360 | 0 |
| instrumental | senior | 420 | 0 |
| instrumental | advanced | 480 | 0 |
| music theatre | pre-junior | 240 | 0 |
| music theatre | junior | 300 | 0 |
| music theatre | intermediate | 360 | 0 |
| music theatre | senior | 420 | 0 |
| music theatre | advanced | 480 | 0 |


## Implementation

From *mfo/config.py* replace move the `DEFAULT_ADJUDICATION_TIME`, `DEFAULT_MOVE_TIME`, and `DEFAULT_FEES` dictionaries, and the `DISCIPLINES` list to a new file in the *test* folder named *qcmf_data.json*. Recast the data in JSON format. While doing this also change the `DEFAULT_ADJUDICATION_TIME` dictionary's data to a nested structure where *discipline* is a key pointing to a dictionary which has *level* as keys. And, change all the values in the `DEFAULT_MOVE_TIME`dictionary to zero. For example:

#### mfo/tests/qcmf_data.json

```
{
    "DEFAULT_ADJUDICATION_TIME": {
        "none": {
            "pre-junior": 240,
            "junior": 300,
            "intermediate": 360,
            "senior": 420,
            "advanced": 480,
        },
        "vocal": {
            ...
        },
        ...
    },
    DEFAULT_MOVE_TIME = {
        "none" : {
            "solo": 0,
            "recital": 0,
            "duet": 0,
            "trio": 0,
            "quartet": 0,
            "quintet": 0,
            "ensemble": 0,
            "composition": 0,
        },
        "vocal" : {
            ...
        },
        ...
    },
    DEFAULT_FEE = {
        "Solo": Decimal("12.00"),
        "Recital": Decimal("16.00"),
        "Duet": Decimal("12.00"),
        "Trio": Decimal("12.00"),
        "Quartet": Decimal("14.00"),
        "Quintet": Decimal("14.00"),
        "Ensemble": Decimal("25.00"),
        "Composition": Decimal("14.00"),
    },
    DISCIPLINES = [
        "Vocal", 
        "Piano", 
        "Organ", 
        "Strings",
        "Recorder",
        "Woodwinds",
        "Brass",
        "Percussion",
        "Instrumental",
        "Musical Theatre",
    ],
    ...
}
```

In addition to the already-existing `DISCIPLINES` list, add one for `TYPES`:

```
    ...
    "TYPES": [
        "Solo",
        "Recital",
        "Duet",
        "Trio",
        "Quartet",
        "Quintet",
        "Ensemble",
        "Composition"
    ],
    ...
```

Also, for levels:

```
    ...
    "LEVELS": [
        "Pre-Junior",
        "Junior",
        "Intermediate",
        "Senior",
        "Advanced"
    ]
    ...
```

#### mfo/config.py

Then, add the location of the test data file in the *config.py* file, next to where we define the test users file:

```
if ENVIRONMENT == "development" or ENVIRONMENT == "lightweight":
    TEST_USERS_FILE = os.environ.get("TEST_USERS_FILE")
    TEST_USERS_FILE = os.path.join(project_dir, TEST_USERS_FILE)
    TEST_DATA_FILE = os.environ.get("TEST_DATA_FILE")
    TEST_USERS_FILE = os.path.join(project_dir, TEST_DATA_FILE)
else:
    TEST_USERS_FILE = os.environ.get("TEST_USERS_FILE")
    TEST_DATA_FILE = os.environ.get("TEST_DATA_FILE")
```

### dotenv files

In the dotenv files, add the filename and path for the *qcmf_data.json* file

#### development.env

```
TEST_DATA_FILE = "/home/blinklet/Projects/Music Festival Program/music-festival-organizer/tests/qcmf_data.json"
```

#### lightweight.env

```
TEST_DATA_FILE = "./tests/qcmf_data.json"
```

#### production.env

```
TEST_DATA_FILE = "./tests/qcmf_data.json"
```

## Command to add class default data to database

Class default data like *discipline*, *fee*, *adjudication time* was previously stored in the *config.py* file and used when importing classes and entries from the Syllabus and the Entries spreadsheet.

Now, I want to place that information in the database, using a Flask CLI command similar to the *test_users* command. After that, I will pull default data from the database instead of the dictionaries defined in the *config.py* file.

### Create new database models

I need to create new database models. 

* **Discipline**: define each discipline. It may be that, in the future, I allow different festivals to set their own disciplines
  * id
  * discipline
  * description
* **PerformanceType**: define each type. May also be changeable in the future.
  * id
  * type
  * description
* **Level**: define each level. May also be changeable in the future.
  * id
  * level
  * description
* **DefaultTime**: for each discipline, type, and level, set a time. Primary key is time/discipline/type/level
  * id
  * discipline: Foreign key to Discipline ID
  * type: Foreign key to Type ID
  * level: Foreign key to Level ID
  * time_type: "adjudication" or "move"
  * time: integer (time in seconds)

#### mfo/database/models.py

I created the models as follows in the *models.py* module:

```
class Discipline(db.Model):
    __tablename__ = 'disciplines'
    
    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
    name: Mapped[str] = mapped_column(nullable=False)
    description: Mapped[Optional[str]] = mapped_column(nullable=True)

    default_times: Mapped[List["DefaultTimes"]] = relationship("DefaultTimes", back_populates="discipline")

    __table_args__ = (
        UniqueConstraint('name', name='discipline_uc'),
    )


class PerformanceType(db.Model):
    __tablename__ = 'performance_types'
    
    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
    name: Mapped[str] = mapped_column(nullable=False)
    description: Mapped[Optional[str]] = mapped_column(nullable=True)

    default_times: Mapped[List["DefaultTimes"]] = relationship("DefaultTimes", back_populates="performance_type")
    default_fees: Mapped[List["DefaultFee"]] = relationship("DefaultFee", back_populates="performance_type")
    __table_args__ = (
        UniqueConstraint('name', name='performance_type_uc'),
    )


class Level(db.Model):
    __tablename__ = 'levels'
    
    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)
    name: Mapped[str] = mapped_column(nullable=False)
    description: Mapped[Optional[str]] = mapped_column(nullable=True)

    default_times: Mapped[List["DefaultTimes"]] = relationship("DefaultTimes", back_populates="level")

    __table_args__ = (
        UniqueConstraint('name', name='level_uc'),
    )


class DefaultTimes(db.Model):
    __tablename__ = 'default_times'
    
    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)

    discipline_id: Mapped[int] = mapped_column(ForeignKey('disciplines.id'), nullable=False)
    performance_type_id: Mapped[Optional[int]] = mapped_column(ForeignKey('performance_types.id'), nullable=True)
    level_id: Mapped[Optional[int]] = mapped_column(ForeignKey('levels.id'), nullable=True)
    time_type: Mapped[str] = mapped_column(nullable=False)
    time: Mapped[int] = mapped_column(nullable=False)

    discipline: Mapped["Discipline"] = relationship("Discipline", back_populates="default_times")
    performance_type: Mapped["PerformanceType"] = relationship("PerformanceType", back_populates="default_times")
    level: Mapped["Level"] = relationship("Level", back_populates="default_times")

    __table_args__ = (
        UniqueConstraint('discipline_id', 'performance_type_id', 'level_id', 'time_type', name='default_times_uc'),
    )

class DefaultFee(db.Model):
    __tablename__ = 'default_fees'
    
    id: Mapped[int] = mapped_column(primary_key=True, autoincrement=True)

    performance_type_id = mapped_column(ForeignKey('performance_types.id'), nullable=False)
    performance_type = relationship("PerformanceType", back_populates="default_fees")

    fee: Mapped[Decimal] = mapped_column(DECIMAL(precision=8, scale=2), nullable=False)

    __table_args__ = (
        UniqueConstraint('performance_type_id', name='default_fees_uc'),
    )
```

> **NOTE:** At this point, I could recast the other tables, like FestivalClass and Repertoire, so that, instead of hard-coding *discipline* and *type*, they use foreign keys and relationships to point to the appropriate row in the *Discipline* and *PerformanceType* tables. I will do that in a future PR.

### Command logic

In the *mfo/database.commands.py* module, add a new function that reads in the test data file and populates the new database tables:

#### mfo/database.commands.py

The following *test_data* command will be run using the flask CLI. Run it after you run the *create* and *test_users* commands.

```
# ...
from sqlalchemy import select
from sqlalchemy.sql import func
from decimal import Decimal

# ...
from mfo.database.models import Profile, Discipline, PerformanceType, Level, DefaultTimes, DefaultFee

# ...

@bp.cli.command('test_data')
@flask.cli.with_appcontext
def test_users():
    try:
        data_file = flask.current_app.config['TEST_DATA_FILE']
    except KeyError:
        print("TEST_DATA_FILE environment variable is not defined")
        return
    
    data_file_path = Path(data_file)
    if os.path.isfile(data_file_path):
        pass
        print("found TEST_DATA_FILE file")
    else:
        print("file defined in TEST_DATA_FILE environment variable does not exist")
        return

    with data_file_path.open() as source:
        data = json.load(source)

    for item in data:
        print(item)

    for discipline in data['DISCIPLINES']:
        # CHECK IF DISCIPLINE EXISTS using select statement
        if db.session.execute(
            select(Discipline)
            .where(func.lower(Discipline.name) == discipline.lower())
        ).scalar():
            print(f"Discipline {discipline} already exists. Skipping this discipline.")
            continue
        discipline_entry = Discipline()
        discipline_entry.name = discipline
        discipline_entry.description = ""
        db.session.add(discipline_entry)
    db.session.commit()

    for class_type in data['TYPES']:
        if db.session.execute(
            select(PerformanceType)
            .where(func.lower(PerformanceType.name) == class_type.lower())
        ).scalar():
            print(f"PerformanceType {class_type} already exists. Skipping this PerformanceType.")
            continue
        performance_type = PerformanceType()
        performance_type.name = class_type
        performance_type.description = ""
        db.session.add(performance_type)
    db.session.commit()

    for level in data['LEVELS']:
        if db.session.execute(select(Level).where(Level.name == level)).scalar():
            print(f"Level {level} already exists. Skipping this Level.")
            continue
        level_entry = Level(name=level, description="")
        db.session.add(level_entry)
    db.session.commit()

    for type, fee in data['DEFAULT_FEE'].items():
        if db.session.execute(
                select(DefaultFee)
                .where(func.lower(PerformanceType.name) == type.lower())
                .join(DefaultFee.performance_type)
            ).scalar():
            print(f"DefaultFee for {type} already exists. Skipping this DefaultFee.")
            continue

        performance_type = db.session.execute(
            select(PerformanceType).where(PerformanceType.name == type)
        ).scalar()
        if performance_type:
            default_fee = DefaultFee(
                performance_type=performance_type,
                fee=Decimal(fee)
            )
            db.session.add(default_fee)
    db.session.commit()

    for discipline, levels in data['DEFAULT_ADJUDICATION_TIME'].items():
        discipline_entry = db.session.execute(
            select(Discipline)
            .where(func.lower(Discipline.name) == discipline.lower())
        ).scalar()

        for level, time in levels.items():
            level_entry = db.session.execute(
                select(Level)
                .where(func.lower(Level.name) == level.lower())
            ).scalar()
            if discipline_entry and level_entry:
                if db.session.execute(
                        select(DefaultTimes)
…